cog-allocation-system/requirements.txt
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
PyYAML>=6.0